Chetty et al. (2014) Commuting Zone Data

Format

A data frame with 707 observations and 39 variables:

state

State abbreviation

cz

Commuting zone identifier

s_1

Latitude (decimal degrees)

s_2

Longitude (decimal degrees)

am

Absolute upward mobility

fracblack

Fraction black

racseg

Racial segregation

segpov25

Segregation of poverty

fraccom15

Fraction commuting < 15 min

hipc

Household income per capita

gini

Gini coefficient

incsh1

Income share top 1\ tsrTeacher-student ratio tspercTest score percentile hsdropHigh school dropout rate scindSocial capital index fracrelFraction religious crimerCrime rate fracsmFraction single mothers fracdivFraction divorced fracmarFraction married loctrLocal tax rate colpcCollege population share coltuiCollege tuition colgradCollege graduation rate manshareManufacturing share chimpChange in imports tlfprTotal labor force participation rate migirateMigration inflow rate migorateMigration outflow rate fracforFraction foreign-born

Chetty, R., Hendren, N., Kline, P., and Saez, E. (2014). "Where is the Land of Opportunity? The Geography of Intergenerational Mobility in the United States." Quarterly Journal of Economics, 129(4), 1553-1623.Data sourced from the SPUR Stata package at https://github.com/pdavidboll/SPUR. chetty Data on intergenerational mobility and its correlates across US commuting zones, used in Chetty, Hendren, Kline, and Saez (2014). Adapted from the SPUR Stata package example data. datasets